+++ src/declaration.d.ts
@@ -7,3 +7,7 @@ declare module '*.scss' {
   const content: Record<string, string>;
   export default content;
 }
+
+interface Window {
+  writer: any;
+}